# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH 26128 Livestock Surveillance API",
    "description": "Checkpoint 3 livestock-health reporting, human review, privacy-preserving surveillance, and verified-only model governance API. This is a decision-support prototype, not a diagnostic system or clinically validated deployment.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "alerts and outbox"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_v1_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Alerts Api V1 Alerts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/outbox/process": {
      "post": {
        "tags": [
          "alerts and outbox"
        ],
        "summary": "Process Outbox",
        "operationId": "process_outbox_api_v1_alerts_outbox_process_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Process Outbox Api V1 Alerts Outbox Process Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/acknowledge": {
      "post": {
        "tags": [
          "alerts and outbox"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_alerts__alert_id__acknowledge_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Acknowledge Api V1 Alerts  Alert Id  Acknowledge Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mlops/overview": {
      "get": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Overview",
        "operationId": "overview_api_v1_mlops_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Overview Api V1 Mlops Overview Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mlops/candidates/{candidate_id}/review": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Review Candidate",
        "operationId": "review_candidate_api_v1_mlops_candidates__candidate_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Review Candidate Api V1 Mlops Candidates  Candidate Id  Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mlops/datasets": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Build Dataset",
        "operationId": "build_dataset_api_v1_mlops_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetBuildIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Build Dataset Api V1 Mlops Datasets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mlops/models/evaluate": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_api_v1_mlops_models_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEvaluationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluate Api V1 Mlops Models Evaluate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mlops/models/{model_id}/approval": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Approve",
        "operationId": "approve_api_v1_mlops_models__model_id__approval_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromotionApprovalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Api V1 Mlops Models  Model Id  Approval Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mlops/models/{model_id}/promote": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Promote",
        "operationId": "promote_api_v1_mlops_models__model_id__promote_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Promote Api V1 Mlops Models  Model Id  Promote Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mlops/models/{model_id}/rollback": {
      "post": {
        "tags": [
          "verified active learning"
        ],
        "summary": "Rollback",
        "operationId": "rollback_api_v1_mlops_models__model_id__rollback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Rollback Api V1 Mlops Models  Model Id  Rollback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcknowledgeIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AcknowledgeIn"
      },
...
      "CandidateEvaluationIn": {
        "properties": {
          "dataset_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Version Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 4,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "BASELINE_EQUIVALENT",
              "INTENTIONAL_REGRESSION_FIXTURE"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_version_id",
          "name",
          "mode"
        ],
        "title": "CandidateEvaluationIn"
      },
      "CandidateReviewIn": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "note"
        ],
        "title": "CandidateReviewIn"
      },
      "DatasetBuildIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 4,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Name"
          },
          "allow_small_synthetic_demo": {
            "type": "boolean",
            "title": "Allow Small Synthetic Demo",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DatasetBuildIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromotionApprovalIn": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "title": "Decision"
          },
          "rationale": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 4,
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "rationale"
        ],
        "title": "PromotionApprovalIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}